+++ src/views/table.js
@@ -1,24 +1,33 @@
 /* eslint-env browser */
 
-const hasOwnProperty = Object.prototype.hasOwnProperty;
-const self = data => data;
+import { createElement } from '../core/utils/dom.js';
+const hasOwnProperty = Object.hasOwnProperty;
 
 function configFromName(name) {
     return {
         header: name,
         view: 'table-cell',
-        data: obj => obj[name]
+        data: obj => obj[name],
+        sorting: `$[${JSON.stringify(name)}] asc`
     };
 }
 
 function resolveColConfig(name, config) {
-    return typeof config === 'string'
-        ? {
-            data: self,
+    if (typeof config === 'string') {
+        return {
+            ...configFromName(name),
             content: config
+        };
+    }
+
+    return hasOwnProperty.call(config, 'content') || hasOwnProperty.call(config, 'data')
+        ? {
+            header: name,
+            view: 'table-cell',
+            ...config
         }
         : {
-            data: hasOwnProperty.call(config, 'content') ? self : obj => obj[name],
+            ...configFromName(name),
             ...config
         };
 }
@@ -28,17 +37,28 @@ export default function(discovery) {
         let { cols, rowConfig, limit } = config;
         let colsMap = cols && typeof cols === 'object' ? cols : {};
         let scalarCol = false;
+        let currentSortingCol = null;
+        let currentSortingColAsc = null;
 
         if (!Array.isArray(data)) {
             data = data ? [data] : [];
         }
 
-        const headEl = el.appendChild(document.createElement('thead'));
-        const bodyEl = el.appendChild(document.createElement('tbody'));
+        const headEl = el.appendChild(createElement('thead'));
+        const bodyEl = el.appendChild(createElement('tbody'));
         const moreEl = el
-            .appendChild(document.createElement('tbody'))
-            .appendChild(document.createElement('tr'))
-            .appendChild(document.createElement('td'));
+            .appendChild(createElement('tbody'))
+            .appendChild(createElement('tr'))
+            .appendChild(createElement('td'));
+
+        const render = (orderedData) => {
+            bodyEl.innerHTML = '';
+            moreEl.innerHTML = '';
+            discovery.view.renderList(bodyEl, this.composeConfig({
+                view: 'table-row',
+                cols
+            }, rowConfig), orderedData, context, 0, discovery.view.listLimit(limit, 25), moreEl);
+        };
 
         if (Array.isArray(cols)) {
             cols = cols.map((def, idx) =>
@@ -47,7 +67,6 @@ export default function(discovery) {
                     : {
                         header: 'col' + idx,
                         view: 'table-cell',
-                        data: self,
                         ...def
                     }
             );
@@ -73,17 +92,16 @@ export default function(discovery) {
                 cols.push({
                     header: '[value]',
                     view: 'table-cell',
-                    data: data => String(data)
+                    data: String
                 });
             }
 
             colNames.forEach(name =>
-                cols.push({
-                    ...configFromName(name),
-                    ...hasOwnProperty.call(colsMap, name)
+                cols.push(
+                    hasOwnProperty.call(colsMap, name)
                         ? resolveColConfig(name, colsMap[name])
-                        : null
-                })
+                        : configFromName(name)
+                )
             );
         } else {
             console.warn('config.cols and data is not an array, no way to build a table');
@@ -103,15 +121,41 @@ export default function(discovery) {
                     discovery.queryBool(whenData, data, context) ? { content } : undefined;
             }
 
-            headEl.appendChild(createElement('th'))
-                .textContent = col.header;
+            const headerCellEl = headEl.appendChild(createElement('th'));
+
+            headerCellEl.textContent = col.header;
+
+            const sorting = discovery.query(col.sorting, null, context);
+            if (typeof sorting === 'function') {
+                headerCellEl.classList.add('sortable');
+                headerCellEl.addEventListener('click', () => {
+                    if (currentSortingCol === headerCellEl) {
+                        if (currentSortingColAsc) {
+                            headerCellEl.classList.remove('asc');
+                            headerCellEl.classList.add('desc');
+                            currentSortingColAsc = false;
+                            render(data.slice().sort((a, b) => -sorting(a, b)));
+                        } else {
+                            headerCellEl.classList.remove('desc');
+                            currentSortingCol = null;
+                            render(data);
+                        }
+                    } else {
+                        if (currentSortingCol !== null) {
+                            currentSortingCol.classList.remove('asc', 'desc');
+                        }
+
+                        currentSortingCol = headerCellEl;
+                        currentSortingColAsc = true;
+                        headerCellEl.classList.add('asc');
+                        render(data.slice().sort(sorting));
+                    }
+                });
+            }
         }
 
         moreEl.colSpan = cols.length;
-        discovery.view.renderList(bodyEl, this.composeConfig({
-            view: 'table-row',
-            cols
-        }, rowConfig), data, context, 0, discovery.view.listLimit(limit, 25), moreEl);
+        render(data);
     }, {
         tag: 'table'
     });